requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
contourpy==1.3.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
Flask==2.0.1
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
gotrue==2.8.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
imageio==2.35.1
importlib-metadata==4.6.4
ipython==8.12.3
itsdangerous==2.1.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
netifaces==0.11.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.0
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
pooch==1.8.2
postgrest==0.16.11
prompt_toolkit==3.0.48
protobuf==5.28.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.5.2
Pygments==2.18.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
PyMatting==1.1.12
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.1
PyYAML==5.4.1
pyzmq==26.2.0
realtime==2.0.2
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.1
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
sympy==1.13.3
systemd-python==234
tifffile==2024.9.20
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==2.0.2
yarg==0.1.9
yarl==1.11.1
zipp==1.0.0
